+++ src/components/GameCard.tsx
@@ -1,4 +1,3 @@
-import { Game } from '../interfaces/Game';
 import PlatformIconList from './PlatformIconList';
 import CriticScore from './CriticScore';
 import imageCropper from '../services/image-cropper';
@@ -7,6 +6,8 @@ import { Card, CardBody, HStack, Heading, Image } from '@chakra-ui/react';
 import Rating from './Rating';
 
 import { Link } from 'react-router-dom';
+
+import Game from '../interfaces/Game';
 
 interface Props {
   game: Game;
