+++ src/components/Navbar.tsx
@@ -1,7 +1,6 @@
 import { HStack, Text } from '@chakra-ui/react';
 import ColorModeSwitch from './ColorModeSwitch';
 import SearchInput from './SearchInput';
-import { Link } from 'react-router-dom';
 
 export default function Navbar() {
   return (
@@ -14,7 +13,7 @@ export default function Navbar() {
         letterSpacing={2}
         border={2}
       >
-        <Link to="/">R A W G</Link>
+        <a href="/">R A W G</a>
       </Text>
       <SearchInput />
       <ColorModeSwitch />
