+++ src/hooks/useGame.ts
@@ -1,6 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
 import APIClient from '../services/api-client';
-import { Game } from '../interfaces/Game';
+
+import Game from '../interfaces/Game';
 
 const apiClient = new APIClient<Game>('/games');
