+++ src/hooks/useGames.ts
@@ -3,7 +3,8 @@ import APIClient, { FetchResponse } from '../services/api-client';
 
 import ms from 'ms';
 import useGameQueryStore from '../store';
-import { Game } from '../interfaces/Game';
+
+import Game from '../interfaces/Game';
 
 const apiClient = new APIClient<Game>('/games');
