+++ src/hooks/usePlatforms.ts
@@ -3,7 +3,8 @@ import platforms from '../data/platforms';
 import { useQuery } from '@tanstack/react-query';
 import APIClient from '../services/api-client';
 import ms from 'ms';
-import { Platform } from '../interfaces/Platform';
+
+import Platform from '../interfaces/Platform';
 
 const apiClient = new APIClient<Platform>('/platforms/lists/parents');
