+++ src/hooks/useTrailers.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import APIClient from '../services/api-client';
-import { Trailer } from '../interfaces/Trailer';
+import Trailer from '../interfaces/Trailer';
 
 const useTrailers = (gameId: number) => {
   const apiClient = new APIClient<Trailer>(`/games/${gameId}/movies`);
