+++ src/interfaces/Genre.ts
@@ -1,4 +1,4 @@
-export interface Genre {
+export default interface Genre {
   id: number;
   name: string;
   image_background: string;
