+++ src/interfaces/Platform.ts
@@ -1,4 +1,4 @@
-export interface Platform {
+export default interface Platform {
   id: number;
   name: string;
   slug: string;
