+++ src/interfaces/Publisher.ts
@@ -1,4 +1,4 @@
-export interface Publisher {
+export default interface Publisher {
   id: number;
   name: string;
 }
