+++ src/interfaces/ScreenShots.ts
@@ -1,4 +1,4 @@
-export interface ScreenShots {
+export default interface ScreenShots {
   id: number;
   image: string;
   width: number;
